Refactor models and training package topology
