+++ dfode_kit/dfode_core/train/formation.py
@@ -1,9 +1,5 @@
-import cantera as ct
-import numpy as np
+"""Compatibility shim for :mod:`dfode_kit.training.formation`."""
 
-def formation_calculate(mechanism):
-    gas = ct.Solution(mechanism) 
-    gas.TPY = 298.15, ct.one_atm,'O2:1' 
-    partial_molar_enthalpy = gas.partial_molar_enthalpies/gas.molecular_weights
-    print(partial_molar_enthalpy)
-    return partial_molar_enthalpy
+from dfode_kit.training.formation import formation_calculate
+
+__all__ = ["formation_calculate"]
