+++ dfode_kit/dfode_core/train/registry.py
@@ -1,29 +1,10 @@
-from __future__ import annotations
+"""Compatibility shim for :mod:`dfode_kit.training.registry`."""
 
-from typing import Callable, Dict
+from dfode_kit.training.registry import create_trainer, get_trainer_factory, register_trainer, registered_trainers
 
-
-TrainerFactory = Callable[..., object]
-_TRAINER_REGISTRY: Dict[str, TrainerFactory] = {}
-
-
-def register_trainer(name: str, factory: TrainerFactory) -> None:
-    if not name:
-        raise ValueError("Trainer name must be non-empty.")
-    _TRAINER_REGISTRY[name] = factory
-
-
-def get_trainer_factory(name: str) -> TrainerFactory:
-    try:
-        return _TRAINER_REGISTRY[name]
-    except KeyError as exc:
-        available = ", ".join(sorted(_TRAINER_REGISTRY)) or "<none>"
-        raise KeyError(f"Unknown trainer '{name}'. Available trainers: {available}") from exc
-
-
-def create_trainer(name: str, **kwargs):
-    return get_trainer_factory(name)(**kwargs)
-
-
-def registered_trainers():
-    return tuple(sorted(_TRAINER_REGISTRY))
+__all__ = [
+    "create_trainer",
+    "get_trainer_factory",
+    "register_trainer",
+    "registered_trainers",
+]
