+++ dfode_kit/training/__init__.py
@@ -1,4 +1,4 @@
-"""Compatibility shims for :mod:`dfode_kit.training`."""
+"""Training configuration, registries, and execution helpers."""
 
 from importlib import import_module
 
@@ -8,13 +8,13 @@ __all__ = [
     "OptimizerConfig",
     "TrainerConfig",
     "TrainingConfig",
-    "create_trainer",
     "default_training_config",
+    "with_overrides",
+    "create_trainer",
     "get_trainer_factory",
     "register_trainer",
     "registered_trainers",
     "train",
-    "with_overrides",
 ]
 
 _ATTRIBUTE_MODULES = {
@@ -22,19 +22,19 @@ _ATTRIBUTE_MODULES = {
     "OptimizerConfig": ("dfode_kit.training.config", "OptimizerConfig"),
     "TrainerConfig": ("dfode_kit.training.config", "TrainerConfig"),
     "TrainingConfig": ("dfode_kit.training.config", "TrainingConfig"),
-    "create_trainer": ("dfode_kit.training.registry", "create_trainer"),
     "default_training_config": ("dfode_kit.training.config", "default_training_config"),
+    "with_overrides": ("dfode_kit.training.config", "with_overrides"),
+    "create_trainer": ("dfode_kit.training.registry", "create_trainer"),
     "get_trainer_factory": ("dfode_kit.training.registry", "get_trainer_factory"),
     "register_trainer": ("dfode_kit.training.registry", "register_trainer"),
     "registered_trainers": ("dfode_kit.training.registry", "registered_trainers"),
     "train": ("dfode_kit.training.train", "train"),
-    "with_overrides": ("dfode_kit.training.config", "with_overrides"),
 }
 
 
 def __getattr__(name):
     if name not in _ATTRIBUTE_MODULES:
-        raise AttributeError(f"module 'dfode_kit.dfode_core.train' has no attribute '{name}'")
+        raise AttributeError(f"module 'dfode_kit.training' has no attribute '{name}'")
 
     module_name, attribute_name = _ATTRIBUTE_MODULES[name]
     module = import_module(module_name)
