+++ docs/agents/train-config-plan.md
@@ -1,10 +1,10 @@
 # Training/config refactor plan
 
 ## Scope
-Keep changes tightly limited to `dfode_kit/dfode_core/{model,train}` and the train CLI surface needed to select registered components. Do not redesign data loading, labeling, or DeepFlame integration in this slice.
+Keep changes tightly limited to the model/training package boundary (`dfode_kit/models`, `dfode_kit/training`, plus compatibility shims under `dfode_kit/dfode_core/{model,train}`) and the train CLI surface needed to select registered components. Do not redesign data loading, labeling, or DeepFlame integration in this slice.
 
 ## Current audit
-- `dfode_kit/dfode_core/train/train.py` hard-codes:
+- `dfode_kit/training/train.py` hard-codes:
   - model architecture (`MLP([2+n_species, 400, 400, 400, 400, n_species-1])`)
   - optimizer (`Adam`)
   - loop hyperparameters (`max_epochs`, LR schedule, batch size)
